+++ src/com/musicott/task/ImportTask.java
@@ -19,18 +19,28 @@
 package com.musicott.task;
 
 import java.io.File;
+import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
+
+import org.jaudiotagger.audio.exceptions.CannotReadException;
+import org.jaudiotagger.audio.exceptions.InvalidAudioFrameException;
+import org.jaudiotagger.audio.exceptions.ReadOnlyFileException;
+import org.jaudiotagger.tag.TagException;
 
 import javafx.application.Platform;
 import javafx.concurrent.Task;
 
+import com.mpatric.mp3agic.InvalidDataException;
+import com.mpatric.mp3agic.UnsupportedTagException;
 import com.musicott.SceneManager;
 import com.musicott.error.ErrorHandler;
 import com.musicott.error.ParseException;
 import com.musicott.model.Track;
 import com.musicott.task.parser.FlacParser;
+import com.musicott.task.parser.M4aParser;
 import com.musicott.task.parser.Mp3Parser;
+import com.musicott.task.parser.WavParser;
 
 /**
  * @author Octavio Calleya
@@ -80,7 +90,7 @@ public class ImportTask extends Task<List<Track>>{
 	@Override
 	protected void cancelled() {
 		super.cancelled();
-		updateMessage("Canceled");
+		updateMessage("Cancelled");
 	}
 	
 	private void scanFolder(File folder) {
@@ -100,21 +110,23 @@ public class ImportTask extends Task<List<Track>>{
 						else
 							if(m4a && file.getName().substring(file.getName().length()-3).equals("m4a")) {
 								updateProgress(++currentFiles, numFiles);
-								//TODO M4aParser
+								list.add(M4aParser.parseM4a(file));
 							}
 							else
 								if(wav && file.getName().substring(file.getName().length()-3).equals("wav")) {
 									updateProgress(++currentFiles, numFiles);
-									//TODO WavParser
+									list.add(WavParser.parseWavFile(file));
 								}
 								else
 									if(flac && file.getName().substring(file.getName().length()-4).equals("flac")) {
 										updateProgress(++currentFiles, numFiles);
 										list.add(FlacParser.parseFlacFile(file));
 									}
-					} catch (Exception e) {
+					} catch (CannotReadException | IOException | TagException | ReadOnlyFileException | InvalidAudioFrameException | UnsupportedTagException |InvalidDataException e) {
 						ParseException pe = new ParseException("Parsing Error", e, file);
 						ErrorHandler.getInstance().addParseException(pe);
+					} catch (ParseException e) {
+						ErrorHandler.getInstance().addParseException(e);
 					}
 				}
 	}