+++ src/com/musicott/task/parser/M4aParser.java
@@ -30,24 +30,28 @@ import org.jaudiotagger.audio.AudioFileIO;
 import org.jaudiotagger.audio.exceptions.CannotReadException;
 import org.jaudiotagger.audio.exceptions.InvalidAudioFrameException;
 import org.jaudiotagger.audio.exceptions.ReadOnlyFileException;
-import org.jaudiotagger.tag.FieldKey;
 import org.jaudiotagger.tag.TagException;
-import org.jaudiotagger.tag.flac.FlacTag;
+import org.jaudiotagger.tag.mp4.Mp4FieldKey;
+import org.jaudiotagger.tag.mp4.Mp4Tag;
 
+import com.musicott.error.ParseException;
 import com.musicott.model.Track;
 
 /**
  * @author Octavio Calleya
  *
  */
-public class FlacParser {
+public class M4aParser {
 
 	private static Track track;
 	
-	public static Track parseFlacFile(final File fileToParse) throws CannotReadException, IOException, TagException, ReadOnlyFileException, InvalidAudioFrameException {
+	public static Track parseM4a(final File fileToParse) throws CannotReadException, IOException, TagException, ReadOnlyFileException, InvalidAudioFrameException, ParseException {
 		track = new Track();
 		AudioFile audioFile = AudioFileIO.read(fileToParse);
-		FlacTag tag = (FlacTag) audioFile.getTag();
+		String encoding = audioFile.getAudioHeader().getEncodingType();
+		if(!encoding.equals("AAC"))
+			throw new ParseException(encoding+" encoding not supported", fileToParse);
+		Mp4Tag tag = (Mp4Tag) audioFile.getTag();
 		track.setFileFolder(new File(fileToParse.getParent()).getAbsolutePath());
 		track.setFileName(fileToParse.getName());
 		track.setInDisk(true);
@@ -55,51 +59,63 @@ public class FlacParser {
 		track.setBitRate(Integer.parseInt(audioFile.getAudioHeader().getBitRate()));
 		track.setTotalTime(Duration.seconds(audioFile.getAudioHeader().getTrackLength()));
 		checkCover(tag);
-		for(FieldKey t: FieldKey.values()) {
-			switch (t){
+		for(Mp4FieldKey fk: Mp4FieldKey.values()) {
+			switch(fk) {
 				case TITLE:
-					track.setName(tag.getFirst(t));
+					track.setName(tag.getFirst(fk));
 					break;
 				case ARTIST:
-					track.setArtist(tag.getFirst(t));
+					track.setArtist(tag.getFirst(fk));
 					break;
 				case ALBUM:
-					track.setAlbum(tag.getFirst(t));
+					track.setAlbum(tag.getFirst(fk));
 					break;
 				case ALBUM_ARTIST:
-					track.setAlbumArtist(tag.getFirst(t));
+					track.setAlbumArtist(tag.getFirst(fk));
 					break;
 				case BPM:
 					try {
-						track.setBpm(Integer.parseInt(tag.getFirst(t)));
-					} catch (NumberFormatException e) {}
+						track.setBpm(Integer.parseInt(tag.getFirst(fk)));
+					}
+					catch (NumberFormatException e) {}
 					break;
 				case COMMENT:
-					track.setComments(tag.getFirst(t));
+					track.setComments(tag.getFirst(fk));
 					break;
 				case GENRE:
-					track.setGenre(tag.getFirst(t));
+					if(track.getGenre().equals(""))
+						track.setGenre(tag.getFirst(fk));
+					break;
+				case GENRE_CUSTOM:
+					if(track.getGenre().equals(""))
+						track.setGenre(tag.getFirst(fk));
 					break;
 				case GROUPING:
-					track.setLabel(tag.getFirst(t));
+					track.setLabel(tag.getFirst(fk));
 					break;
-				case IS_COMPILATION:
-					track.setCompilation(tag.getFirst(t).equals("true") ? Boolean.valueOf(tag.getFirst(t)) : false);
+				case COMPILATION:
+					try {
+						track.setCompilation(Integer.parseInt(tag.getFirst(fk)) == 1 ? true : false);
+					}
+					catch (NumberFormatException e) {}
 					break;
 				case TRACK:
 					try {
-						track.setTrackNumber(Integer.parseInt(tag.getFirst(t)));
-					} catch (NumberFormatException e) {}
+						track.setTrackNumber(Integer.parseInt(tag.getFirst(fk)));
+					}
+					catch (NumberFormatException e) {}
 					break;
-				case DISC_NO:
+				case DISCNUMBER:
 					try {
-						track.setDiscNumber(Integer.parseInt(tag.getFirst(t)));
-					} catch (NumberFormatException e) {}
+						track.setDiscNumber(Integer.parseInt(tag.getFirst(fk)));
+					}
+					catch (NumberFormatException e) {}
 					break;
-				case YEAR:
+				case MM_ORIGINAL_YEAR:
 					try {
-						track.setYear(Integer.parseInt(tag.getFirst(t)));
-					} catch (NumberFormatException e) {}
+						track.setYear(Integer.parseInt(tag.getFirst(fk)));
+					}
+					catch (NumberFormatException e) {}
 					break;
 				default:
 			}
@@ -107,7 +123,7 @@ public class FlacParser {
 		return track;
 	}
 	
-	private static void checkCover(FlacTag tag) {
+	public static void checkCover(Mp4Tag tag) {
 		if(!tag.getArtworkList().isEmpty())
 			track.setHasCover(true);
 		else {